requirements.txt
streamlit
pyngrok
langchain
groq
pydantic
langchain-pinecone
pinecone-client
langchain-core
bitsandbytes
transformers
accelerate
langchain-huggingface
langchain-community
pypdf
langgraph
langchain-text-splitters
langchain-groq
torch